imessage-bots/src/bots/meeting-scheduler/requirements.txt
fastapi[standard]>=0.115.13,<0.116.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
openai>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
python-dateutil>=2.8.0 